refactored evaluation code to be object-oriented
